Remove redundant packages
