+++ internal/frontend/hegel/README.md
@@ -1,0 +1,3 @@
+The sources in this package were copied with the intent of converting them to a true frontend
+in the future. They've been modified to compile so they don't interfere with builds but they aren't
+used anywhere.
